+++ src/main/java/com/cgr/base/infrastructure/persistence/entity/UserEntity.java
@@ -1,6 +1,5 @@
 package com.cgr.base.infrastructure.persistence.entity;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -15,7 +14,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import lombok.Data;
@@ -50,19 +48,13 @@ public class UserEntity {
 
     private String cargo;
 
+    private String userType;
+
     @ManyToMany
     @JsonIgnoreProperties({ "users", "handler", "hibernateLazyInitializer" })
     @JoinTable(name = "users_roles", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"), uniqueConstraints = {
             @UniqueConstraint(columnNames = { "user_id", "role_id" }) })
     private List<RoleEntity> roles;
-
-    @OneToMany(mappedBy = "user")
-    private List<LogEntity> logs = new ArrayList<>();
-
-    public void addRol(RoleEntity roleEntity) {
-        this.roles.add(roleEntity);
-        roleEntity.getUsers().add(this);
-    }
 
     public void mapActiveDirectoryUser(UserEntity userAD) {
         this.fullName = userAD.getFullName();
