+++ src/main/java/com/cgr/base/infrastructure/persistence/entity/contracts/ContractExecutionTracking.java
@@ -1,7 +1,6 @@
-package com.cgr.base.infrastructure.persistence.entity;
+package com.cgr.base.infrastructure.persistence.entity.contracts;
 
 import java.util.Date;
-import java.util.List;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,7 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
@@ -18,37 +16,38 @@ import lombok.Data;
 
 @Entity
 @Data
-@Table(name = "Supports")
-public class Support {
+@Table(name = "ContractExecutionTracking")
+public class ContractExecutionTracking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int supportId;
+    private int trackingId;
 
     @ManyToOne
     @JoinColumn(name = "contractId")
     private Contract contract;
 
-    @Column(nullable = false)
-    private int supportType;
-
-    @Column(nullable = false, length = 65533)
-    private String description;
-
     @Temporal(TemporalType.DATE)
-    private Date issueDate;
+    private Date trackingDate;
 
-    @Column(nullable = false, length = 65533)
-    private String digitalFile;
+    @Column(nullable = false)
+    private double physicalProgress;
 
-    @Column(nullable = false, length = 250)
-    private String issueResponsible;
+    @Column(nullable = false)
+    private double financialProgress;
 
     @Column(nullable = false, length = 30)
-    private String supportStatus;
+    private String contractStatus;
 
-    @OneToMany(mappedBy = "support")
-    private List<SupportHistory> histories;
+    @Column(nullable = false, length = 65533)
+    private String complianceIndicators;
+
+    @Column(nullable = false, length = 65533)
+    private String observations;
+
+    @Column(nullable = false, length = 250)
+    private String trackingResponsible;
 
     // Getters and Setters
 }
+
