+++ src/main/java/com/cgr/base/infrastructure/persistence/entity/contracts/Entitys.java
@@ -1,4 +1,4 @@
-package com.cgr.base.infrastructure.persistence.entity;
+package com.cgr.base.infrastructure.persistence.entity.contracts;
 
 import java.util.List;
