+++ src/main/java/com/cgr/base/infrastructure/persistence/entity/contracts/Support.java
@@ -1,6 +1,7 @@
-package com.cgr.base.infrastructure.persistence.entity;
+package com.cgr.base.infrastructure.persistence.entity.contracts;
 
 import java.util.Date;
+import java.util.List;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,6 +10,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
@@ -16,28 +18,37 @@ import lombok.Data;
 
 @Entity
 @Data
-@Table(name = "SupportHistory")
-public class SupportHistory {
+@Table(name = "Supports")
+public class Support {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int historyId;
+    private int supportId;
 
     @ManyToOne
-    @JoinColumn(name = "supportId")
-    private Support support;
+    @JoinColumn(name = "contractId")
+    private Contract contract;
+
+    @Column(nullable = false)
+    private int supportType;
 
     @Column(nullable = false, length = 65533)
-    private String performedAction;
+    private String description;
 
     @Temporal(TemporalType.DATE)
-    private Date actionDate;
-
-    @Column(nullable = false, length = 250)
-    private String responsibleUser;
+    private Date issueDate;
 
     @Column(nullable = false, length = 65533)
-    private String comments;
+    private String digitalFile;
+
+    @Column(nullable = false, length = 250)
+    private String issueResponsible;
+
+    @Column(nullable = false, length = 30)
+    private String supportStatus;
+
+    @OneToMany(mappedBy = "support")
+    private List<SupportHistory> histories;
 
     // Getters and Setters
 }
