--- src/main/java/com/cgr/base/infrastructure/persistence/repository/logs/ILogsRepositoryJpa.java
@@ -1,11 +1,0 @@
-package com.cgr.base.infrastructure.persistence.repository.logs;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.cgr.base.infrastructure.persistence.entity.LogEntity;
-
-@Repository
-public interface ILogsRepositoryJpa extends JpaRepository<LogEntity, Long> {
-    
-}
